+++ src/main/java/edu/ie3/datamodel/io/naming/FieldNamingStrategy.java
@@ -87,6 +87,13 @@ public class FieldNamingStrategy {
   public static final String X_SC_B = "xScB";
   public static final String X_SC_C = "xScC";
 
+  // cable types
+  public static final String LIMIT_TEMPERATURE = "limitTemperature";
+  public static final String FREQUENCY = "frequency";
+  public static final String ELECTRICAL_CAPACITANCE = "electricalCapacitance";
+  public static final String THERMAL_RESISTIVITY = "thermalResistivity";
+  public static final String THERMAL_CAPACITANCE = "thermalCapacitance";
+
   // efficiency
   public static final String ETA = "eta";
   public static final String ETA_CONV = "etaConv";