+++ src/test/groovy/edu/ie3/datamodel/io/factory/input/participant/WecInputFactoryTest.groovy
@@ -36,10 +36,10 @@ class WecInputFactoryTest extends Specification implements FactoryTestHelper {
     given: "a system participant input type factory and model data"
     def inputFactory = new WecInputFactory()
     Map<String, String> parameter = [
-      "uuid"            : "91ec3bcf-1777-4d38-af67-0bf7c9fa73c7",
-      "operatesfrom"    : "",
-      "operatesuntil"   : "2019-01-01T00:00:00+01:00[Europe/Berlin]",
-      "id"              : "TestID",
+      "uuid" : "91ec3bcf-1777-4d38-af67-0bf7c9fa73c7",
+      "operatesfrom" : "",
+      "operatesuntil" : "2019-01-01T00:00:00+01:00[Europe/Berlin]",
+      "id" : "TestID",
       "qcharacteristics": "cosPhiFixed:{(0.0,1.0)}"
     ]
     def inputClass = WecInput